+++ crates/stats-agg/src/stats2d.rs
@@ -6,7 +6,7 @@ use twofloat::TwoFloat;
 
 mod stats2d_flat_serialize;
 
-#[derive(Debug, PartialEq, Copy, Clone, Serialize, Deserialize)]
+#[derive(Debug, PartialEq, Eq, Copy, Clone, Serialize, Deserialize)]
 #[repr(C)]
 pub struct StatsSummary2D<T: FloatLike> {
     pub n: u64, // count